add product order order_product

+++ app/controllers/users_controller.ts
@@ -1,4 +1,5 @@
 
+import UserPolicy from '#policies/user_policy';
 import UsersService from '#services/user_service';
 import { inject } from '@adonisjs/core';
 import type { HttpContext } from '@adonisjs/core/http'
@@ -6,19 +7,29 @@ import type { HttpContext } from '@adonisjs/core/http'
 @inject()
 export default class UsersController {
   constructor(private readonly userService: UsersService){}
-  /**
-   * Display a list of resource
-   */
-  async index({pagination}: HttpContext) {
+  async index({bouncer, response,pagination}: HttpContext) {
     try {
+      if (await bouncer.with(UserPolicy).denies('index')) {
+        return response.unauthorized({message: 'Unauthorized'})
+      }
       return await this.userService.index(pagination)
     } catch (error) {
       throw error
     }
   }
+
   async show({ params }: HttpContext) {
     try {
       return await this.userService.findById(params.id)
+    } catch(error){
+      throw error
+    }
+  }
+
+  async verifyUser({ request }: HttpContext) {
+    try {
+      const {email, otp} = request.only(['email', 'otp'])
+      return await this.userService.verify(otp, email)
     } catch(error){
       throw error
     }
@@ -40,8 +51,11 @@ export default class UsersController {
     }
   }
 
-  async destroy({ params }: HttpContext) {
+  async destroy({bouncer, response, params }: HttpContext) {
     try {
+      if (await bouncer.with(UserPolicy).denies('destroy')) {
+        return response.unauthorized({message: 'Unauthorized'})
+      }
       return await this.userService.destroy(params.id)
     } catch (error) {
       throw error
